bootstrap logic corrected.

+++ crontab-scripts/run.sh
@@ -22,17 +22,22 @@ if [ ! -z "$pid" ]; then
     [ ! -z "$pid" ] && rm -f "$LOCK"
 fi
 
-BOOTSTRAP_REQUIRED="$HOME/bootstrap_required"
-if [ -f "$BOOTSTRAP_REQUIRED" ]; then
-    [ -x ./bootstrap.sh ] && ./bootstrap.sh
-    rm -f "$BOOTSTRAP_REQUIRED"
+npaths=$(ls -1d *-automated-ci 2>/dev/null | wc -l)
+path="$PWD"
+scriptwd="$PWD"
+if [ "$npaths" == "1" -a -d *-automated-ci ]; then
+    path=${path}/*-automated-ci
+    cd $path
 fi
 
-
-npaths=$(ls -1d *-automated-ci 2>/dev/null | wc -l)
-if [ "$npaths" == "1" -a -d *-automated-ci ]; then
-    path=*-automated-ci
-    cd $path
+BOOTSTRAP_REQUIRED="$HOME/bootstrap_required"
+if [ -f "$BOOTSTRAP_REQUIRED" ]; then
+    if [ "$PWD" == $(cat "$BOOTSTRAP_REQUIRED") ]; then
+        cd $scriptwd
+        [ -x ./bootstrap.sh ] && ./bootstrap.sh
+        rm -f "$BOOTSTRAP_REQUIRED"
+        cd $path
+    fi
 fi
 
 echo $$ > "$LOCK"
